requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ansicon==1.89.0
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.1.6
django-crispy-forms==2.0
django-environ==0.10.0
django-picklefield==3.1
django-q==1.3.9
frozenlist==1.3.3
idna==3.4
jinxed==1.2.0
multidict==6.0.4
pymemcache==4.0.0
python-dateutil==2.8.2
redis==3.5.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
wcwidth==0.2.6
yarl==1.8.2
